remove meta class for fitter
Fix test since fitter meta class has been removed
